BLD: Remove CUDA requirements from recipe
Using a CUDA enabled version of PtychoNN is not strictly necessary. Persons using PtychoNN on apple silicon or whatever should be allowed to run Ptychodus will all the features even if PtychoNN is slower. If PtychoNN does require the CUDA version of pytorch, then it should export that requirement itself.